sunulek-api/requirements.txt
# Django
Django>=5.0,<6.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0

# Database
psycopg2-binary>=2.9.9

# CORS
django-cors-headers>=4.3.0

# Environment variables
python-decouple>=3.8

# API Documentation
drf-spectacular>=0.27.0

# Images
Pillow>=10.0.0

# Production
gunicorn>=21.0.0
whitenoise>=6.6.0

# Email
django-anymail>=10.0

# Filtering
django-filter>=23.0

# Development
black>=23.0.0
flake8>=6.0.0
